### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Refactoring

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
Yes

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
Port 4 tests to helio_distributed and clean up the code to make them pass.

----
#### AI description  (iteration 1)
#### PR Classification
Porting initial tests to a distributed environment.

#### PR Summary
This pull request ports the first four tests to a distributed environment, ensuring that the query evaluation functions work correctly in this setup.
- Added distributed test cases for various BSON expression evaluations in `src/test/regress/expected/bson_expr_eval_tests.out`.

+++ pg_helio_api/src/pg_helio_api.c
@@ -22,12 +22,20 @@ void _PG_init(void);
 void _PG_fini(void);
 
 
+bool SkipHelioApiLoad = false;
+
+
 /*
  * _PG_init gets called when the extension is loaded.
  */
 void
 _PG_init(void)
 {
+	if (SkipHelioApiLoad)
+	{
+		return;
+	}
+
 	if (!process_shared_preload_libraries_in_progress)
 	{
 		ereport(ERROR, (errmsg(
@@ -56,5 +64,10 @@ _PG_init(void)
 void
 _PG_fini(void)
 {
+	if (SkipHelioApiLoad)
+	{
+		return;
+	}
+
 	UninstallHelioApiPostgresHooks();
 }
